Refactor results parsers
